git-svn-id: https://cvs.dai-labor.de/jiac/toolipse3/trunk@33472 d8bac29e-b7f5-0310-9bdb-a06bcefdf3b8

+++ de.dailab.vsdt.trafo.jiacbeans/src/de/dailab/vsdt/trafo/jiacbeans/export/JiacBeansResultSaver.java
@@ -1,16 +1,23 @@
 package de.dailab.vsdt.trafo.jiacbeans.export;
 
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
+
+import javax.swing.JDialog;
+import javax.swing.JOptionPane;
 
 import jiacbeans.AgentBean;
 
 import org.eclipse.emf.codegen.jet.JETEmitter;
+import org.eclipse.emf.codegen.merge.java.JControlModel;
+import org.eclipse.emf.codegen.merge.java.JMerger;
+import org.eclipse.emf.codegen.merge.java.facade.ast.ASTFacadeHelper;
 
 import de.dailab.vsdt.trafo.MappingResultSaver;
 import de.dailab.vsdt.trafo.jiacbeans.export.generated.JavaCodeGenerator;
-import de.dailab.vsdt.trafo.jiacbeans.util.JavaCode;
 
 public class JiacBeansResultSaver extends MappingResultSaver {
 
@@ -21,6 +28,11 @@ public class JiacBeansResultSaver extends MappingResultSaver {
 
 	@Override
 	protected boolean internalSave(File baseDirectory) throws IOException {
+		JControlModel model = new JControlModel(); 
+		ASTFacadeHelper astFacadeHelper = new ASTFacadeHelper(); 
+		model.initialize(astFacadeHelper, "mergerules.xml"); 
+		JMerger jMerger = new JMerger(model); 
+		
 		JiacBeansExportWrapper wrapper = (JiacBeansExportWrapper) this.wrapper;
 		wrapper.setBaseDirectory(baseDirectory);
 		JavaCodeGenerator generator = new JavaCodeGenerator();
@@ -39,9 +51,26 @@ public class JiacBeansResultSaver extends MappingResultSaver {
 			if(!folder.exists())folder.mkdirs();
 			String fileName = bean.getName();
 			File f = new File(folder,fileName+".java");
-			FileWriter writer = new FileWriter(f);
-			writer.write(generator.generate(bean));
-			writer.flush();
+			String content = generator.generate(bean);
+			if(!f.exists()){
+				FileWriter writer = new FileWriter(f);
+				writer.write(content);
+				writer.flush();
+			}else{
+				//source=new generated code 
+				try { //try to merge
+					jMerger.setSourceCompilationUnit(jMerger.createCompilationUnitForContents(content));
+					jMerger.setTargetCompilationUnit(jMerger.createCompilationUnitForInputStream(new FileInputStream(f))); //target=last generated code
+					jMerger.merge(); 
+					String result = jMerger.getTargetCompilationUnit().getContents(); 
+					System.out.println(result);
+				}catch (Exception e) {
+					//if something goes wrong overwrite the file
+					FileWriter writer = new FileWriter(f);
+					writer.write(content);
+					writer.flush();
+				} 
+			}
 		}
 		return true;
 	}
